Reorganizar un poco los tests
