+++ test/models/caja/caja_movimientos_test.rb
@@ -1,48 +1,10 @@
 # encoding: utf-8
 require 'test_helper'
 
-class CajaTest < ActiveSupport::TestCase
+class CajaMovimientosTest < ActiveSupport::TestCase
   setup do
     # Caja sin movimientos para la mayoría de los tests
     @caja = create :caja
-  end
-
-  test 'es válida' do
-    [ :build, :build_stubbed, :create].each do |metodo|
-      assert_valid_factory metodo, :caja
-      assert_valid_factory metodo, :chequera
-      assert_valid_factory metodo, :banco
-    end
-  end
-
-  test 'totaliza en pesos por default' do
-    2.times { create :movimiento, caja: @caja, monto: Money.new(1000) }
-
-    assert_equal Money.new(2000), @caja.total
-  end
-
-  test 'totaliza por moneda' do
-    2.times { create :movimiento, caja: @caja, monto: Money.new(1000) }
-    2.times { create :movimiento, caja: @caja, monto: Money.new(500, 'USD') }
-
-    assert_equal Money.new(2000), @caja.total
-    assert_equal Money.new(1000, 'USD'), @caja.total('USD')
-  end
-
-  test 'todos los totales' do
-    2.times { create :movimiento, caja: @caja, monto: Money.new(1000) }
-    2.times { create :movimiento, caja: @caja, monto: Money.new(500, 'USD') }
-    create :movimiento, caja: @caja, monto: Money.new(500, 'EUR')
-
-    assert_equal ({ 'ARS' => Money.new(2000),
-                    'USD' => Money.new(1000, 'USD'),
-                    'EUR' => Money.new(500, 'EUR') }), @caja.totales
-  end
-
-  test 'caja en cero devuelve money' do
-    assert @caja.movimientos.empty?
-    assert_equal Money.new(0), @caja.total
-    assert_equal ({ 'ARS' => Money.new(0) }), @caja.totales
   end
 
   test 'extrae si alcanza y devuelve el movimiento' do
@@ -143,15 +105,6 @@ class CajaTest < ActiveSupport::TestCase
     assert_nil @recibo_interno
   end
 
-  test 'unifica los tipos prefiriendo el existente' do
-    tipo_existente = 'Cajón sarasa'
-    create(:caja, tipo: tipo_existente)
-
-    assert_equal tipo_existente, create(:caja, tipo: ' Cajón    sarasa ').tipo
-    assert_equal tipo_existente, create(:caja, tipo: 'Cajon sarasa').tipo
-    assert_equal tipo_existente, create(:caja, tipo: 'cajón sarasa').tipo
-  end
-
   test 'transferir dineros de una caja a otra' do
     origen = create :caja, :con_fondos
     destino = create :caja
@@ -164,17 +117,5 @@ class CajaTest < ActiveSupport::TestCase
     assert_equal 2, recibo.movimientos.count
     assert_equal dineros, destino.total
     assert_equal 0, origen.total
-  end
-
-  test 'no permite tipos iguales en una misma obra y con el mismo numero' do
-    caja1 = create :caja, tipo: 'Personal', obra_id: '1234', numero: ''
-
-    # no permite cajas con mismo tipo
-    assert_raise ActiveRecord::RecordInvalid do
-      create :caja, tipo: 'Personal', obra_id: '1234', numero: ''
-    end
-
-    # a menos que tengan numeros diferentes
-    assert create :caja, tipo: 'Personal', obra_id: '1234', numero:'1'
   end
 end
